Refactor MembersPage and enhance member management
Refactored MembersPage to use a GridView layout for improved UI/UX, while preserving the old ListView layout in MembersOldPage. Added new properties (`IsCaptain`, `AccountName`) to MemberObservable for richer member data. Introduced a confirmation dialog for member removal in MembersViewModel. Updated localized strings and telemetry logging for better user feedback and debugging. Refreshed image assets and simplified MembersPage.xaml.cs structure.

+++ AscendiaApp/ViewModels/MembersViewModel.cs
@@ -8,7 +8,9 @@ using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using LCTWorks.Telemetry;
 using LCTWorks.WinUI.Dialogs;
+using LCTWorks.WinUI.Extensions;
 using Microsoft.Extensions.Logging;
+using Microsoft.UI.Xaml.Controls;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -166,7 +168,19 @@ public partial class MembersViewModel : ObservableObject
         {
             return;
         }
+        var results = await _dialogService.ShowDialogAsync(new ContentDialog
+        {
+            Title = "Dialog-RemoveMemberTitle".GetTextLocalized(),
+            Content = "Dialog-RemoveMemberMessage".GetTextLocalized(),
+            PrimaryButtonText = "Dialog-RemoveMemberPrimaryButtonText".GetTextLocalized(),
+            SecondaryButtonText = "Dialog-RemoveMemberSecondaryButtonText".GetTextLocalized(),
+        });
+        if (results != ContentDialogResult.Primary)
+        {
+            return;
+        }
         IsLoading = true;
+
         var success = await _communityService.RemoveMemberAsync(SelectedMember.Record.Id!);
         if (success)
         {
